requirements.txt


# fastapi
# uvicorn[standard]
sentence-transformers
faiss-cpu
numpy
pdfplumber
python-multipart
# pydantic
python-telegram-bot==13.15
pdfplumber
sentence-transformers
faiss-cpu
numpy
fastapi
pydantic
requests
python-telegram-bot==13.15
# openai
openai==1.3.0
python-multipart==0.0.6
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
textblob==0.17.1
vaderSentiment==3.3.2


